requirements.txt
blinker==1.9.0
chardet==5.2.0
click==8.1.8
Flask==2.3.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
python-docx==0.8.11
reportlab==4.3.1
tabulate==0.9.0
Werkzeug==2.3.7
